reorganize sites directory
