+++ job/steps/collaborative_filtering/save_defaults.py
@@ -10,7 +10,7 @@ from db.mappings.model import ModelType
 from db.mappings.recommendation import Rec
 from job.helpers.datetime import time_decay
 from lib.config import config
-from sites.site import Site
+from sites.templates.site import Site
 
 MAX_RECS = config.get("MAX_RECS")
