+++ job/steps/collaborative_filtering/warehouse.py
@@ -11,7 +11,7 @@ import s3fs
 from job.helpers.datetime import chunk_name
 from lib.config import config
 from lib.events import PING_INTERVAL, Event
-from sites.site import Site
+from sites.templates.site import Site
 
 
 class Table(Enum):
