+++ sites/helpers/singleton.py
@@ -11,7 +11,8 @@ class SingletonABCMeta(ABCMeta):
     Inheriting from `ABCMeta` ensures behavior of both a singleton, defined here, and an abstract base
     class, defined by `ABCMeta`.
 
-    For why we decided to use the singleton pattern here, visit https://github.com/LocalAtBrown/article-rec-training-job/pull/149#discussion_r968885680
+    For the reason that we decided to use the singleton pattern here, visit
+    https://github.com/LocalAtBrown/article-rec-training-job/pull/149#discussion_r968885680
     """
 
     _instances = {}
